Refactor the time feature.
