+++ src/main/kotlin/com/dignicate/p30a/controller/di/Module.kt
@@ -1,7 +1,7 @@
 package com.dignicate.p30a.controller.di
 
 import com.dignicate.p30a.controller.automobile.AutomobileController
-import com.dignicate.p30a.controller.currenttime.CurrentTimeController
+import com.dignicate.p30a.controller.time.CurrentTimeController
 import org.koin.dsl.module
