+++ src/main/kotlin/com/dignicate/p30a/controller/time/TimeRoutes.kt
@@ -1,4 +1,4 @@
-package com.dignicate.p30a.controller.currenttime
+package com.dignicate.p30a.controller.time
 
 import com.dignicate.p30a.controller.ErrorResponse
 import io.ktor.http.HttpStatusCode
@@ -8,8 +8,8 @@ import io.ktor.server.routing.Route
 import io.ktor.server.routing.get
 import org.koin.java.KoinJavaComponent.getKoin
 
-fun Route.currentTimeRoutes() {
-    get("/current-time") {
+fun Route.timeRoutes() {
+    get("/time/v1/current") {
         val controller: CurrentTimeController = getKoin().get()
         try {
             call.respond(HttpStatusCode.OK, controller.getCurrentTime())
