+++ src/main/kotlin/com/dignicate/p30a/domain/di/Module.kt
@@ -1,7 +1,7 @@
 package com.dignicate.p30a.domain.di
 
 import com.dignicate.p30a.domain.automobile.GetCompaniesUseCase
-import com.dignicate.p30a.domain.currenttime.GetCurrentTimeUseCase
+import com.dignicate.p30a.domain.time.GetCurrentTimeUseCase
 import org.koin.dsl.module
