+++ src/main/kotlin/com/dignicate/p30a/domain/time/CurrentTimeDomainObject.kt
@@ -1,4 +1,4 @@
-package com.dignicate.p30a.domain.currenttime
+package com.dignicate.p30a.domain.time
 
 import java.time.Instant
