+++ src/main/kotlin/com/dignicate/p30a/domain/time/CurrentTimeRepository.kt
@@ -1,4 +1,4 @@
-package com.dignicate.p30a.domain.currenttime
+package com.dignicate.p30a.domain.time
 
 interface CurrentTimeRepository {
     suspend fun getCurrentTime(): Result<CurrentTime>
