+++ src/main/kotlin/com/dignicate/p30a/domain/time/GetCurrentTimeUseCase.kt
@@ -1,4 +1,4 @@
-package com.dignicate.p30a.domain.currenttime
+package com.dignicate.p30a.domain.time
 
 class GetCurrentTimeUseCase(
     private val repository: CurrentTimeRepository
