+++ src/main/kotlin/com/dignicate/p30a/plugins/Routing.kt
@@ -2,7 +2,7 @@ package com.dignicate.p30a.plugins
 
 import com.dignicate.p30a.controller.automobile.automobileRoutes
 import com.dignicate.p30a.controller.common.commonRoutes
-import com.dignicate.p30a.controller.currenttime.currentTimeRoutes
+import com.dignicate.p30a.controller.time.timeRoutes
 import io.ktor.server.application.*
 import io.ktor.server.plugins.swagger.*
 import io.ktor.server.routing.*
@@ -11,7 +11,7 @@ fun Application.configureRouting() {
     routing {
         commonRoutes()
         automobileRoutes()
-        currentTimeRoutes()
+        timeRoutes()
         swaggerUI(path = "swagger", swaggerFile = "openapi/documentation.yaml")
     }
 }
